Legacy rp update

+++ geonode/contrib/datatables/views.py
@@ -21,14 +21,18 @@ from geonode.contrib.datatables.forms import JoinTargetForm,\
 #from geonode.contrib.dataverse_connect.layer_metadata import LayerMetadata        # object with layer metadata
 from shared_dataverse_information.worldmap_datatables.forms import MapLatLngLayerRequestForm
 
-
-from geonode.contrib.msg_util import msg, msgt, msgn, msgx
+from geonode.contrib.msg_util import *
 
 from .models import DataTable, JoinTarget, TableJoin, LatLngTableMappingRecord
-from .utils import create_point_col_from_lat_lon,\
-    standardize_name,\
-    attempt_tablejoin_from_request_params,\
+from geonode.contrib.datatables.name_helper import standardize_column_name
+
+from geonode.contrib.datatables.utils import attempt_tablejoin_from_request_params,\
     attempt_datatable_upload_from_request_params
+#from geonode.contrib.datatables.utils_joins import drop_view_from_table_join
+from geonode.contrib.datatables.utils_lat_lng import create_point_col_from_lat_lon
+from geonode.contrib.datatables.db_helper import get_datastore_connection_string
+
+from geonode.contrib.datatables.column_checker import ColumnHelper
 
 logger = logging.getLogger(__name__)
 
@@ -219,8 +223,19 @@ def tablejoin_remove(request, tj_id):
         json_msg = MessageHelperJSON.get_json_fail_msg(err_msg)
         return HttpResponse(json_msg, mimetype='application/json', status=401)
 
+
     # -----------------------------------------------
-    # Delete it!
+    # Delete the view...the TableJoin "Layer"
+    # -----------------------------------------------
+    """
+    # Not needed!
+    view_dropped, err_msg = drop_view_from_table_join(tj)
+    if not view_dropped:
+        json_msg = MessageHelperJSON.get_json_fail_msg(err_msg)
+        return HttpResponse(json_msg, mimetype='application/json', status=400)
+    """
+    # -----------------------------------------------
+    # Delete The DataTable, JoinLayer, and TableJoin objects!
     # -----------------------------------------------
     try:
         tj.datatable.delete()
@@ -298,11 +313,31 @@ def datatable_upload_and_join_api(request):
         json_msg = MessageHelperJSON.get_json_fail_msg(err_msg, data_dict=f.errors)
         return HttpResponse(json_msg, mimetype="application/json", status=400)
 
+    # ----------------------------------------------------
+    # Does the DataTable join column need to be char?
+    #  - Check if the existing target join column is char?
+    # ----------------------------------------------------
+    (check_worked, force_char_convert) = ColumnHelper.is_char_column_conversion_recommended(\
+                f.cleaned_data['layer_name'],\
+                f.cleaned_data['layer_attribute'])
+    if not check_worked:
+        err_msg = 'Could not check the target column type'
+        logger.error(err_msg)
+        json_msg = MessageHelperJSON.get_json_fail_msg(err_msg)
+        return HttpResponse(json_msg, mimetype="application/json", status=400)
+
+    if force_char_convert:  # It is, make sure the Datatable col will be char
+        force_char_column = join_props['table_attribute']
+    else:                   # Nope, let the datatable col stand, good or bad
+        force_char_column = None
 
     # ----------------------------------------------------
     # Create a DataTable object from the file
     # ----------------------------------------------------
-    (success, data_table_or_error) = attempt_datatable_upload_from_request_params(request, request.user)
+    (success, data_table_or_error) = attempt_datatable_upload_from_request_params(\
+                                request,\
+                                request.user,\
+                                force_char_column=force_char_column)
     if not success:
         json_msg = MessageHelperJSON.get_json_fail_msg(data_table_or_error)
         return HttpResponse(json_msg, mimetype="application/json", status=400)
@@ -316,7 +351,7 @@ def datatable_upload_and_join_api(request):
     #
     join_props['table_name'] = data_table_or_error.table_name
     original_table_attribute = join_props['table_attribute']
-    sanitized_table_attribute = standardize_name(original_table_attribute)
+    sanitized_table_attribute = standardize_column_name(original_table_attribute)
     join_props['table_attribute'] = sanitized_table_attribute
 
     (success, tablejoin_obj_or_err_msg) = attempt_tablejoin_from_request_params(join_props, request.user)
@@ -365,7 +400,6 @@ def datatable_upload_lat_lon_api(request):
     #   Set the new table/layer owner
     #
     new_table_owner = request.user
-    msg('datatable_upload_lat_lon_api 1')
 
     # --------------------------------------
     # (1) Datatable Upload
